Various updates
Updated variable formatting  `___variableName` is now `_variableName`
Added the functionality of `always on top`. When an element is actively being dragged, it will relayer over everything else so it can be seen.
Renamed a few params for readability.
Removed some code and restructured some code
Removed the excessive calls to `World.getGameSize()` and just stored it's value since its static.
Removed the code from the event function `onDragEnd` and `onMove` and it's now handled internally.
Children elements now call the event functions if their parent is being dragged. (Only if defined)
The `pX` and `pY` params in the event functions now represent the new position of the element.
All elements now have a `interfaceName` variable. It refers to the interface this element belongs to.
All elements now have access to the variable `defaultDisplay` it holds the default layer and plane of this element.
All elements now have access to the variable `defaultSize` it holds the default width and height of this element.
All elements now have access to the variable `defaultPos` it holds the default xPos and yPos of this element.

Introduced a preprocessor `#DEFINE MAX_PLANE` to handle the `always on top` functionality.

Docs for these updates coming soon 🚧