added a new isMousedDown function to interfaces
This function belongs to all interface elements, and is useful for checking if a mouse event originated from this event before allowing it to continue. Such as `element.onMouseClick` gets called even if the `element.onMouseDown` event did not start on that element, which can feel a little weird depending on what feel you're going for.