Update aInterfaceUtils.vs
Small optimizations: no more creating a new object for each new element being dragged, now reuses the same object.
Added a new variable to the client to track the state of dragging. The client's new variable `dragging` is true if the client is dragging something around. It is false when dragging stops. This can be useful for things like preventing `onMouseEnter` events from firing if the client is dragging something and the mouse wanders onto a diob. (That shouldn't count as a real `onMouseEnter` so this variable can help with that.

Added a new variable to interface elements named `defaultScreenPercentage` which is just prep for a future update that will auto position diobs that have `preventAutoScale` enabled when the screen is resized. This is a read only variable.