refactor-comment-maintenance
Added a comment to explain why the code was written a certain way that differed from the lib's standard.
Some code in this library executes before the Client object is created. Therefore it needs to be explained that the code is attached to the Type rather than the Client object. (As it looks rather confusing at first glance). (And made me want to fix it)

Renamed some variables for readability.  (Some of these variables were named weirdly in the past, and did not cleanly express what their purpose was.

Added some sanity checks for event functions called via this library. Since these are not built in functions, the developer can assign a value to a variable that is not callable. This sanity check, checks if the event function is of a function's type.