update-refactor
Created a function for the client called `releaseElement` so it's easy to release an element from being dragged from any other part of the code.
Added two new Client event functions:

Client.onMouseEnter // this is called when the mouse re-enters the game window
Client.onMouseLeave // this is called when the mouse leaves the game window

When dragging an interface element and the mouse leaves the screen, the element is released now.

Removed the parameters from event functions. (They were pretty useless and not needed).