update-refactor
Added a sanity check for the `enter` and `leave` functions. This prevents it from throwing errors when the game isn't loaded.
Changed some references from `VS.World.global` to `VS.Client`. It's shorter and the best way to reference the library when applicable.
When mousing over an interface that can be dragged, the mouse cursor will now change to the `move` cursor. 
This library performs smart checks to see if the aInventory library is being used in tandem with this library. If it is, it does not allow the cursor to change if the user is currently holding a slot inside of the aInventory library.

There is a current bug inside of the Vylocity engine that calls `Client.onMoveMove` whenever you click back on the screen to gain focus. This causes unwanted results to the dragging portion of this library. A polyfill fix has been added until that bug is addressed.

There is a extra sanity check to the isMousedDown function now. If a Diob is being dragged, it is no longer considered to be considered moused down. This is to prevent `click` events from being called when an element that was being dragged is finally dropped.