Update aInterfaceUtils.js
Fixed a few typos.
Did some sanity checks on variables to make sure they had a value and if not defaulted them to 0. As to not get any NaN values.
Added a sanity check inside of a setTimeout to check if the variable still existed, in the chance that the thing self._dragging.element was referencing was deleted and if it was it now resets the data as expected.