Add functionality for elements using preventAutoScale
This update changes `parentElement` from just being a string, to being the actual reference element now. (internally)

This change now allows elements that use preventAutoScale to retain their position on the screen when the window is resized. This works initially for singular elements.

If you want a element to follow its parent's positioning, then it must have parentElement set to be a string containing the name of that parent element. The positioning is based on the % of element on the screen.